This commit was generated by cvs2svn to compensate for changes in r12965, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@12966 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/debug.c
@@ -167,20 +167,20 @@ void debug_maincpu(DWORD reg_pc, CLOCK mclk, const char *dis, BYTE reg_a,
             }  
         }
 
-        log_debug("%04X %ld %02X%02X%02X %s",
-                  reg_pc, (long)mclk, reg_a, reg_x, reg_y, small_dis);
+        log_debug("%04X %ld %02X%02X%02X %s", (unsigned int)reg_pc,
+                  (long)mclk, reg_a, reg_x, reg_y, small_dis);
     } else {
         log_debug(".%04X %03i %03i %10ld  %-20s "
-                  "A=$%02X X=$%02X Y=$%02X SP=$%02X",
-                  reg_pc, RLINE(mclk), RCYCLE(mclk), (long)mclk, dis,
+                  "A=$%02X X=$%02X Y=$%02X SP=$%02X", (unsigned int)reg_pc,
+                  RLINE(mclk), RCYCLE(mclk), (long)mclk, dis,
                   reg_a, reg_x, reg_y, reg_sp);
     }
 }
 
 void debug_drive(DWORD reg_pc, CLOCK mclk, const char *dis, BYTE reg_a)
 {
-    log_debug("Drive: .%04X %10ld %-20s A=$%02x.", reg_pc, (long)mclk, dis,
-              reg_a);
+    log_debug("Drive: .%04X %10ld %-20s A=$%02x.", (unsigned int)reg_pc,
+              (long)mclk, dis, reg_a);
 }
 
 void debug_text(const char *text)
